final fixing

+++ database/migrations/2025_09_22_162345_create_users_table.php
@@ -8,6 +8,9 @@ return new class extends Migration
 {
     public function up(): void
     {
+        if (Schema::hasTable('users')) {
+            return;
+        }
         Schema::create('users', function (Blueprint $table) {
             $table->id();
             $table->string('first_name');
@@ -27,6 +30,9 @@ return new class extends Migration
 
     public function down(): void
     {
+        if (!Schema::hasTable('users')) {
+            return;
+        }
         Schema::dropIfExists('users');
     }
 };
